requirements.txt
boto3
pillow
matplotlib
streamlit